pyproject.toml
[tool.poetry]
name = "tourism"
version = "0.1.0"
description = ""
authors = ["Reza Dalvand <rdalvand@yahoo.com>"]
readme = "README.md"
packages = [{include = "tourism"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
django-split-settings = "^1.2.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
boto3 = "^1.34.5"
django-storages = "^1.14.2"
split-settings = "^1.0.0"
django-cors-headers = "^4.3.1"
kavenegar = "^1.1.2"
djangorestframework-simplejwt = "^5.3.1"



[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
django-environ = "^0.11.2"
sentry-sdk = {extras = ["django"], version = "^1.39.1"}
django-rosetta = "^0.10.0"
django-debug-toolbar = "^4.2.0"
django-rest-swagger = "^2.2.0"
drf-yasg = "^1.21.7"
pytest-django = "^4.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'


[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.django"
python_files = ["test_*.py", "*_test.py", "testing/python/*.py"]
